devrel/requirements.txt
praw>=7.8.0
stackapi>=0.3.0
PyGithub>=2.0.0
aiohttp>=3.9.0
jinja2>=3.1.0
rich>=13.0.0
schedule>=1.2.0
python-dotenv>=1.0.0
